refactor: start modularizing bruteforcer
