+++ src/deps.ts
@@ -1,0 +1,4 @@
+export * as anotherCookieJar from "https://deno.land/x/another_cookiejar@v5.0.3/mod.ts";
+export * as cliffy from "https://deno.land/x/cliffy@v0.25.7/prompt/mod.ts";
+export * as dotenv from "https://deno.land/std@0.201.0/dotenv/mod.ts";
+export * as z from "https://deno.land/x/zod@v3.22.2/mod.ts";
