+++ src/exporters/csv.ts
@@ -1,0 +1,5 @@
+import { ScheduleExporter } from "./types.d.ts";
+
+export default function toCsv(schedule, index) {
+    
+}
