+++ src/exporters/tts.ts
@@ -1,0 +1,5 @@
+import { Schedule } from "../model/schedule.ts";
+
+export function exportSchedule(schedule: Schedule, index: number) {
+
+}
