--- src/parser/search/courses/types.d.ts
@@ -1,6 +1,0 @@
-export type Course = {
-    id: number,
-    code: string,
-    year: number,
-    name: string,
-}
