--- src/parser/search/teachers/types.d.ts
@@ -1,5 +1,0 @@
-export type Teacher = {
-    id: string,
-    name: string,
-    abbreviation: string,
-}
