+++ src/reset.d.ts
@@ -1,0 +1,2 @@
+// Do not add any other lines of code to this file!
+import "npm:@total-typescript/ts-reset";
