+++ src/sigarra/api.ts
@@ -1,0 +1,1 @@
+export function getA
