+++ src/sigarra/fetchers/schedule/teachers/api/mod.ts
@@ -1,0 +1,15 @@
+import { fetchWithAuthentication } from "@/sigarra/fetchers/client.ts";
+import { throwIfNotOk } from "@/sigarra/fetchers/utils.ts";
+import parseSchedule from "@/sigarra/parsers/schedule/api/mod.ts";
+
+export function getTeacherSchedule(teacherId: string, start: string, end: string) {
+    const url = new URL("https://sigarra.up.pt/feup/pt/MOB_HOR_GERAL.DOCENTE");
+    url.searchParams.set("pv_codigo", teacherId);
+    url.searchParams.set("pv_semana_ini", start);
+    url.searchParams.set("pv_semana_fim", end);
+
+    return fetchWithAuthentication(url.toString())
+        .then(throwIfNotOk)
+        .then((response) => response.json())
+        .then(parseSchedule);
+}
