+++ src/sigarra/fetchers/utils.ts
@@ -1,0 +1,7 @@
+export function throwIfNotOk(response: Response) {
+    if (!response.ok) {
+        throw new Error(`Request failed with status ${response.status}`);
+    }
+
+    return response;
+}
